[Sema] Initial commit of -fauto-compile-for-accelerator flag
This commit introduce a new cc1 flag: -fauto-compile-for-accelerator .
Use it with: -Xclang -fauto-compile-for-accelerator .

It modifies several places in Sema module, mostly in function / method overload
resolution logic, by adding additional codes so GPU attribute (CXXRestrictAMPAttr)
would be added in cases normally an error would be diagnosed.

The purpose of this commit is to allow users of HC mode to not have to annotate
functions / methods with [[hc]] attribute. The attribute would be automatically
annotated if a function / method is called by a GPU kernel.

Following cases are considered as of now, which should cover most common cases:

- global function
- static function
- member function of a class
- static member function of a class