[GridLaunch] Move setting "hc_grid_launch" attribute to CodeGenModule::SetLLVMFunctionAttributes
Previously, the string attribute was set in CodeGenModule::
SetLLVMFunctionAttributesForDefinition which prevented external GridLaunch
function declarations to getting said attribute. This meant the
corresponding GridLaunch LLVM passes would not be able to operate on code
where only a forward kernel declaration is present.

The LLVM passes would have also have to be updated accordingly.